extracted incoming date formatter to common package to match outgoing struture
